# lab-data/correction/clients-swagger2.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Référentiel Clients",
    "version": "1.0",
    "description": "Fichier Swagger 2.0 pour une connexion REST dans Informatica Cloud. Remplacer host par l'URL publique du Codespace (sans https://)."
  },
  "host": "VOTRE-CODESPACE-8090.app.github.dev",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "cleApi": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  },
  "security": [
    {
      "cleApi": []
    }
  ],
  "paths": {
    "/v1/clients": {
      "get": {
        "operationId": "listerClients",
        "summary": "Lister les clients, page par page",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 0
          },
          {
            "name": "taille",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 100,
            "minimum": 1,
            "maximum": 100
          },
          {
            "name": "modifieDepuis",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "statut",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "ACTIF",
              "INACTIF",
              "PROSPECT"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Une page de clients",
            "schema": {
              "$ref": "#/definitions/PageClients"
            }
          },
          "401": {
            "description": "Authentification absente ou invalide"
          },
          "429": {
            "description": "Limite de débit atteinte"
          }
        }
      }
    },
    "/v1/clients/{id}": {
      "get": {
        "operationId": "lireClient",
        "summary": "Lire un client",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Le client",
            "schema": {
              "$ref": "#/definitions/Client"
            }
          },
          "404": {
            "description": "Client inexistant"
          }
        }
      }
    }
  },
  "definitions": {
    "PageClients": {
      "type": "object",
      "properties": {
        "donnees": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Client"
          }
        },
        "pagination": {
          "$ref": "#/definitions/Pagination"
        },
        "liens": {
          "$ref": "#/definitions/Liens"
        }
      }
    },
    "Pagination": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "taille": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Liens": {
      "type": "object",
      "properties": {
        "suivant": {
          "type": "string"
        },
        "precedent": {
          "type": "string"
        }
      }
    },
    "Client": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "reference": {
          "type": "string"
        },
        "nom": {
          "type": "string"
        },
        "prenom": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telephone": {
          "type": "string",
          "description": "Peut être null"
        },
        "statut": {
          "type": "string",
          "enum": [
            "ACTIF",
            "INACTIF",
            "PROSPECT"
          ]
        },
        "adresse": {
          "$ref": "#/definitions/Adresse"
        },
        "comptes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Compte"
          }
        },
        "creeLe": {
          "type": "string",
          "format": "date-time"
        },
        "modifieLe": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Adresse": {
      "type": "object",
      "properties": {
        "rue": {
          "type": "string"
        },
        "codePostal": {
          "type": "string"
        },
        "ville": {
          "type": "string"
        },
        "pays": {
          "type": "string"
        }
      }
    },
    "Compte": {
      "type": "object",
      "properties": {
        "numero": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": [
            "COURANT",
            "EPARGNE",
            "TITRES"
          ]
        },
        "solde": {
          "type": "number",
          "format": "double"
        },
        "devise": {
          "type": "string"
        },
        "ouvertLe": {
          "type": "string",
          "format": "date"
        }
      }
    }
  }
}